+++ app/src/main/java/com/aoztg/greengrim/data/local/ChatEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "unread_chat_data")
 data class UnReadChatEntity(
     @PrimaryKey
-    val chatId: Int = 0,
+    val chatId: Long = 0,
     val unReadCount: Int = 0,
     val recentChat: String = "",
     val recentChatTime: String = "",
