+++ app/src/main/java/com/aoztg/greengrim/data/model/response/CreateCertificationResponse.kt
@@ -1,7 +1,7 @@
 package com.aoztg.greengrim.data.model.response
 
 data class CreateCertificationResponse(
-    val certId: Int,
+    val certId: Long,
     val certImg: String,
     val date: String
 )
