+++ app/src/main/java/com/aoztg/greengrim/data/remote/ChatAPI.kt
@@ -12,7 +12,7 @@ interface ChatAPI {
 
     @POST("/visitor/challenges/enter")
     suspend fun enterChat(
-        @Query("id") challengeId: Int
+        @Query("id") challengeId: Long
     ): Response<EnterChatResponse>
 
     @GET("/visitor/challenges/chatrooms")
@@ -20,12 +20,12 @@ interface ChatAPI {
 
     @POST("/visitor/chatrooms/exit")
     suspend fun exitChatRoom(
-        @Query("chatroomId") id: Int
+        @Query("chatroomId") id: Long
     ): Response<Unit>
 
     @GET("/chat/message")
     suspend fun getChatMessage(
-        @Query("roomId") roomId: Int,
+        @Query("roomId") roomId: Long,
         @Query("page") page: Int,
         @Query("size") size: Int
     ): Response<ChatMessageResponse>
